Update table content validation for large tables
* Refs #533 to support reading tables with record counts greater than the max integer value
* Update classes that reference those methods

Other updates to TableDataContentValidationRule to fix other issues that popped up when trying to make this fix, including:
* spaghetti code throughout validateTableDataContents. still lots of code everywhere that should be consolidated, but its a start
* re-use RandomAccessFile and InputStream for data files where there are multiple tables/objects within the same file. there are numerous places throughout the code where we re-read data over and over again. tried to catch some of this, but I'm sure there are more enhancements needed.
* fix record counters. they should be more accurate now.

This commit will only work with the changes from pds4-jparser: NASA-PDS/pds-jparser@2b984cfdfb992fa951f9709d2fdb2b1a684e777f

+++ src/main/java/gov/nasa/pds/tools/validate/rule/pds4/CollectionReferentialIntegrityRule.java
@@ -246,7 +246,7 @@ public class CollectionReferentialIntegrityRule extends AbstractValidationRule {
         }
         entry = reader.getNext();
       }
-      int records = reader.getNumRecords();
+      long records = reader.getNumRecords();
       LOG.debug("getCollectionMembers: collection,numOfCollectionMembers,records {},{},{}",collection,numOfCollectionMembers,records);
 
       if (numOfCollectionMembers>0 && records>0 && numOfCollectionMembers!=records) {
